riscv: first pass at adding ld.aq and st.rl instructions
Per @ZSusskind, we are reusing the `CUSTOM_(0,1)` opcode and matching
the encoding format of `ld.w` and `st.w` instructions to reuse as much
code as possible.

Next steps are figuring out how to write a rewrite pattern to convert
the instruction sequences mentioned in the riscv-isa spec down to the
new `ld.aq` and `st.rl` instructions.