aarch64: first pass at making their atomics codegen pass worse
We want to evaluate how much performance improves by using the
AArch64 `ld.aq` and `st.rl` variants. As a result, we attempt a
rewrite pattern that lowers them to the Memory Data Barrier (MSB)
+ fence combinations that are present in the RISCV spec today.

This will allow us to quantify how much performance is *lost* with the
dumber instructions for the report.